+++ server/.env.dev
@@ -3,3 +3,5 @@ DATABASE_USERNAME=root
 DATABASE_PASSWORD=123456789
 DATABASE_DB=mellowboard
 DATABASE_HOST=127.0.0.1
+ACCESS_SECRET=asdasd
+REFRESH_SECRET=asdasd2
