+++ server/.env.production
@@ -3,3 +3,5 @@ DATABASE_USERNAME=admin
 DATABASE_PASSWORD=mellowboard1234!
 DATABASE_DB=mellowboard
 DATABASE_HOST=mellowboard.cwkgqpjwktns.ap-northeast-2.rds.amazonaws.com
+ACCESS_SECRET=mellowboardaccesssecret!@!@
+REFRESH_SECRET=mellowboardrefreshsecret!@!@
