+++ server/routes/users.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var router = express.Router();
 const users = require('../controllers/users');
-const isValidToken = require("./middleware/isValidToken");
+const isValidToken = require("../middleware/isValidToken");
 
 router.use(isValidToken);
 router.get('/', users.get);
